lib download_protoc: protoc already existing is a warning
This is not an error, unless someone depends on a new version. I'm
not exactly sure how Cargo versions build scripts. I think maybe we
should be downloading protoc to a named file with a version, so we
can detect version upgrades. On the other hand, it is possible that
Cargo's existing versioning mechanism already does this?